enforce single responsibility principle
Divided the KeyboardHook class into two classes as it originally not
only intercepted keyDown events from the system's input loop, but
additionally discriminated them by key type and several keys had their
own, separate event fired when pressed. This was removed in favor of
simplicity - every button fires the same event and it's up to the class
using the KeyboardHook to decide how to treat a given button.
The second class, WordsInterceptor, depends upon KeyboardHook, for now
on concrete implementation, which is subject to change in the future. It
reacts differently to certain ranges of buttons being pressed in order
to construct words typed on the keyboard. A single event it raises is
fired when a single word is finished. It handles well corrections made
during the typing and only returns the word in it's final state.